requirements.txt
aiolimiter==1.2.1
anyio==4.7.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.0.2
pedalboard==0.9.16
pycparser==2.22
pydub==0.25.1
PySoundFile==0.9.0.post1
python-telegram-bot==21.10
sniffio==1.3.1
typing-extensions==4.12.2
werkzeug==3.1.3
zipp==3.21.0
